Store Manager Print Ticket Bug Fix
Disabling of "Print Ticket" button was implemented incorrectly. It allowed for printing tickets on weekdays when the store is closed as long as a single working_hours entry for the store existed.

This bug is fixed.
Redundant and correct check whether the store is open implemented. Conditions are now checked both while accessing the page and while creating an appointment for the proxy customer.